Modification of clear function : Now the function keeps the current background and text color and doesn't use printCharacter function because when we arrive at the last position, the function will shift the screen.
Modification of printCharacter :
The function shift the screen if we arrive at the last position.
All char are move one line to the top and the last line is clear
(but keeps current background and text color)

If we don't keep current background and text color for the clear,
and set the value to 0, we won't be able to see the cursor, not very
practical.